fix(sidebar): apply eligibility filter to watchlist activity cards
